Switch to using UINT (from UINT16) for geometry indices

+++ Samples/Desktop/D3D12Raytracing/src/D3D12RaytracingSimpleLighting/RaytracingHlslCompat.h
@@ -18,7 +18,7 @@
 using namespace DirectX;
 
 // Shader will use byte encoding to access indices.
-typedef UINT16 Index;
+typedef UINT Index;
 #endif
 
 struct SceneConstantBuffer
